Add public in console_io helper.
As it's used in a bundled migration, and is the only way, for now, to get the io in php migration : I think it should be set to public.

IMHO, a cleaner way would be to inject the helper (or even remove the helper to directly use Symfony's Input And Output) in PHPExecutor.